feat: :rocket: builds and template is ready for styling
